+++ extensions/promptlab/src/components/Commands/CommandSetupForm.tsx
@@ -12,7 +12,7 @@ export default function CommandSetupForm(props: {
   commandName: string;
   options: CommandOptions;
   setOptions: React.Dispatch<React.SetStateAction<CommandOptions>>;
-  setCommands?: React.Dispatch<React.SetStateAction<Command[]>>;
+  setCommands?: (commands: Command[]) => void;
 }) {
   const { commandName, options, setOptions, setCommands } = props;
   const [setupFields, setSetupFields] = useState<