+++ extensions/raycast-ollama/CHANGELOG.md
@@ -1,5 +1,9 @@
 # raycast-ollama Changelog
 
+## [BugFix] - 2024-01-24
+
+- Fixed `Cannot read properties of undefined (reading 'split')` caused by API route '/api/show' responding with empty 'parameters' value for some models.
+
 ## [BugFix] - 2024-01-12
 
 - **Command 'Improve Writing'**: fixed error 'Preferred model is not configured in LocalStorage'.